Moved timestamp to helpers
