--- src/utils/currentTimestamp.js
@@ -1,6 +1,0 @@
-// Make unix timestamp
-function getCurrentTimestamp() {
-    return Math.round(new Date().getTime() / 1000)
-}
-
-module.exports = getCurrentTimestamp
